put tinymce in a component for reuse between translation and page form
