+++ angular/app/cms/components/tinymce/cms-tinymce.component.html
@@ -1,0 +1,7 @@
+<div ng-if="vm.options && vm.model != undefined">
+	<textarea
+		ui-tinymce="vm.options"
+		ng-model="vm.model"
+		rows="8"
+	></textarea>
+</div>
